+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Cleanup transcoding temp files
 
+### Fixed
+
+- Use domain from lti jwt token when no consumer site used
+
 ## [5.2.0] - 2024-10-22
 
 ### Added